+++ src/lib.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // SPDX-FileCopyrightText: 2026 matthias
 
-pub mod burn;
+pub mod annot;
 pub mod config;
 pub mod library;
 pub mod tuner;